refactors lock page
