+++ src/utils/LockOverviewUtils.ts
@@ -5,6 +5,7 @@ import {
 } from '@/data/LockConstants'
 import { Result } from '@ethersproject/abi'
 import { BigNumber, ethers } from 'ethers'
+import * as Humanize from 'humanize-plus'
 
 export const calculate4YearsYield = (totalHiiq: number) => {
   let yieldWithA4YearLock = 1 * (1 + 0.75 * 4)
@@ -56,3 +57,34 @@ export const addGasLimitBuffer = (value: BigNumber) =>
   value
     .mul(ethers.BigNumber.from(10000 + 2500))
     .div(ethers.BigNumber.from(10000))
+
+export const formatValue = (value: number) => {
+  if (value !== undefined) {
+    const valueToString = value.toString()
+    return valueToString.length > 6
+      ? Humanize.compactInteger(value, 2)
+      : Humanize.formatNumber(value, 2)
+  }
+  return 0
+}
+
+export const calculateReturn = (
+  userTotalIQLocked: number,
+  lockValue: number,
+  lockEndDate: number | Date,
+  iqToBeLocked: number,
+) => {
+  let tokenToBeLocked = userTotalIQLocked
+  let timeLocked = lockValue
+  if (iqToBeLocked) {
+    tokenToBeLocked += iqToBeLocked
+  }
+  if (typeof lockEndDate !== 'number') {
+    const currentDateTime = new Date().getTime()
+    const lockedTime = lockEndDate.getTime()
+    const differenceInDays = (lockedTime - currentDateTime) / (1000 * 3600 * 24)
+    if (differenceInDays > 0) timeLocked += differenceInDays
+  }
+  const returns = tokenToBeLocked + (tokenToBeLocked * 3 * timeLocked) / 1460
+  return returns
+}
